fix: split files to satisfy file-length lint, drop unused helper
