+++ cmd/server/serve_routes_v1.go
@@ -167,14 +167,10 @@ func setupV1Routes(
 						r.Post("/{agentID}/conversations", conversationHandler.Create)
 						r.Get("/{agentID}/conversations", conversationHandler.List)
 					}
-					// Agent triggers removed — routing is now handled by the
-					// Zira router at /v1/router/triggers.
 				})
-
 				if systemTaskHandler != nil {
 					r.Post("/system/tasks/{taskName}", systemTaskHandler.Run)
 				}
-
 				// Zira Router — unified routing identity for the org.
 				r.Route("/router", func(r chi.Router) {
 					r.Get("/", routerHandler.GetOrCreateRouter)
